style: fix formatting and linting issues
- Fix multi-line match arms formatting in TUI
- Convert struct initializations to consistent multi-line format
- Split .await.context() chains across multiple lines for readability
- Fix clippy warnings: unused variables, unnecessary casts, dead code
- Replace implicit subtraction with saturating_sub for safety
- Collapse nested else-if blocks for cleaner control flow
- Replace vec! with array where appropriate

All changes pass cargo fmt and cargo clippy checks

+++ tests/integration_playback.rs
@@ -170,7 +170,7 @@ fn write_silence_wav(path: &PathBuf) -> std::io::Result<()> {
     let num_samples = sample_rate * duration_secs;
 
     let byte_rate = sample_rate * (bits_per_sample as u32) / 8 * num_channels as u32;
-    let block_align = num_channels * (bits_per_sample / 8) as u16;
+    let block_align = num_channels * (bits_per_sample / 8);
     let subchunk2_size = num_samples * (bits_per_sample as u32 / 8) * num_channels as u32;
     let chunk_size = 36 + subchunk2_size;
